refactor: align hook dapp files structure
