+++ apps/cowswap-frontend/src/modules/hooksStore/dapps/BuildHookApp/styled.tsx
@@ -1,0 +1,60 @@
+import styled from 'styled-components/macro'
+
+export const Wrapper = styled.div`
+  display: flex;
+  flex-flow: column wrap;
+
+  flex-grow: 1;
+`
+
+export const Link = styled.button`
+  border: none;
+  padding: 0;
+  text-decoration: underline;
+  display: text;
+  cursor: pointer;
+  background: none;
+  color: white;
+  margin: 10px 0;
+`
+
+export const Header = styled.div`
+  display: flex;
+  padding: 1.5em;
+
+  p {
+    padding: 0 1em;
+  }
+`
+
+export const ContentWrapper = styled.div`
+  flex-grow: 1;
+  justify-content: center;
+  align-items: center;
+  flex-flow: column wrap;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  padding: 1em;
+  text-align: center;
+`
+
+export const Row = styled.div`
+  display: flex;
+  flex-flow: row wrap;
+  margin: 10px;
+  width: 100%;
+
+  label {
+    margin: 10px;
+    flex-grow: 0;
+    width: 5em;
+  }
+
+  input,
+  textarea {
+    flex-grow: 1;
+  }
+`
