+++ apps/cowswap-frontend/src/modules/hooksStore/dapps/ClaimGnoHookApp/index.tsx
@@ -2,88 +2,18 @@ import { useCallback, useContext, useEffect, useMemo, useState } from 'react'
 
 import gnoLogo from '@cowprotocol/assets/cow-swap/network-gnosis-chain-logo.svg'
 import { GNO } from '@cowprotocol/common-const'
-import { HookDappInternal, HookDappType } from '@cowprotocol/types'
-import { ButtonPrimary, UI } from '@cowprotocol/ui'
+import { HookDapp } from '@cowprotocol/types'
+import { ButtonPrimary } from '@cowprotocol/ui'
 import { BigNumber } from '@ethersproject/bignumber'
 import { CurrencyAmount } from '@uniswap/sdk-core'
 
 import { formatUnits } from 'ethers/lib/utils'
-import styled from 'styled-components/macro'
+
+import { Amount, ContentWrapper, ErrorLabel, Header, Label, Link, LoadingLabel, Wrapper } from './styled'
 
 import { SBC_DEPOSIT_CONTRACT_ADDRESS } from '../../const'
 import { HookDappContext } from '../../context'
 import { useSBCDepositContract } from '../../hooks/useSBCDepositContract'
-
-const TITLE = 'Claim GNO from validators'
-const DESCRIPTION = 'Allows you to withdraw the rewards from your Gnosis validators.'
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-flow: column wrap;
-
-  flex-grow: 1;
-`
-
-const Link = styled.button`
-  border: none;
-  padding: 0;
-  text-decoration: underline;
-  display: text;
-  cursor: pointer;
-  background: none;
-  color: white;
-  margin: 10px 0;
-`
-
-const Header = styled.div`
-  display: flex;
-  padding: 1.5em;
-
-  p {
-    padding: 0 1em;
-  }
-`
-
-const Label = styled.span`
-  color: var(${UI.COLOR_TEXT2});
-`
-
-const ContentWrapper = styled.div`
-  flex-grow: 1;
-  justify-content: center;
-  align-items: center;
-  flex-flow: column wrap;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  padding: 1em;
-  text-align: center;
-`
-
-const Amount = styled.div`
-  font-weight: 600;
-  margin-top: 0.3em;
-`
-
-const ErrorLabel = styled.div`
-  color: var(${UI.COLOR_RED});
-`
-
-const LoadingLabel = styled.div`
-  color: var(${UI.COLOR_TEXT2});
-`
-
-export const PRE_CLAIM_GNO: HookDappInternal = {
-  name: TITLE,
-  description: DESCRIPTION,
-  type: HookDappType.INTERNAL,
-  path: '/hooks-dapps/pre/claim-gno',
-  component: <ClaimGnoHookApp />,
-  image: gnoLogo,
-  version: 'v0.1.1',
-}
 
 /**
  * Dapp that creates the hook to the connected wallet GNO Rewards.
@@ -92,7 +22,7 @@ export const PRE_CLAIM_GNO: HookDappInternal = {
  *    - Proxy: 0x0B98057eA310F4d31F2a452B414647007d1645d9 (https://gnosisscan.io/address/0x0B98057eA310F4d31F2a452B414647007d1645d9#readProxyContract)
  *    - Master: 0x4fef25519256e24a1fc536f7677152da742fe3ef
  */
-export function ClaimGnoHookApp() {
+export function ClaimGnoHookApp({ dapp }: { dapp: HookDapp }) {
   const hookDappContext = useContext(HookDappContext)
   const SbcDepositContract = useSBCDepositContract()
   const [claimable, setClaimable] = useState<BigNumber | undefined>(undefined)
@@ -144,10 +74,10 @@ export function ClaimGnoHookApp() {
           gasLimit: gasLimit.toString(),
           target: SBC_DEPOSIT_CONTRACT_ADDRESS,
         },
-        dapp: PRE_CLAIM_GNO,
+        dapp,
         outputTokens: [CurrencyAmount.fromRawAmount(gno, claimable.toString())],
       },
-      true
+      true,
     )
   }, [callData, gasLimit, hookDappContext, claimable])
 
@@ -166,8 +96,8 @@ export function ClaimGnoHookApp() {
   return (
     <Wrapper>
       <Header>
-        <img src={gnoLogo} alt={TITLE} width="60" />
-        <p>{DESCRIPTION}</p>
+        <img src={gnoLogo} alt={dapp.name} width="60" />
+        <p>{dapp.description}</p>
       </Header>
       <ContentWrapper>
         <ClaimableAmount loading={loading} claimable={claimable} error={error} />
